feat: handled authentication with Socious ID for AuthGuard container
