+++ src/modules/Auth/components/AuthGuardModal/index.types.ts
@@ -1,0 +1,5 @@
+export interface AuthGuardModalProps {
+  open: boolean;
+  handleClose: () => void;
+  onContinue: () => void;
+}
