+++ src/modules/Auth/containers/AuthGuard/index.types.ts
@@ -1,5 +1,3 @@
-import React from 'react';
-
 export type AuthGuardProps = {
   children: React.ReactNode;
   redirectUrl?: string;
